Implement memory efficient multifile pointcloud creation
Iterative addition of pre-created pointclouds was previously used
to create a single pointcloud from multiple .las files. While fast
for small files, it becomes unwieldy with large files as the
constant copying of large arrays becomes expensive.

The new method pre-allocates an array which is filled by iterating
through the files. However, this is fragile as it relies on the
file header providing an accurate report of the number of points.
The headers should be checked first, or a backup routine
implemented, as this could produce subtle bugs.