Store point coordinates as regular array
The `points` structured record array previously used to store
point coordinates was deem inefficient and unecessary. It has
been implemented as a property, with the underlying point storage
changed to a regular 3*n (x, y, z) numpy array, known as `arr`.
This is NOT identical to the previous property `arr`, which was n*3.
This is therefore a breaking change, and very few tests work.

This branch is dedicated to fixing much of code to acknowledge the
pointsectomy.