Implement retiling in 3 dimensions
The `retile` function allows a set of pointclouds to be retiled by
specifying locations in x, y and/or z dimensions which form the
boundaries between tiles, producing a 3D numpy array.

This function comes with a lot of extra associated functions, and
there is more to come. It is likely that these will be factored out
into a seperate retile module at some point in the future.